Refactor: Integrate Home page to Main Page
